requirements.txt
edge-tts==7.0.0
openai==1.70.0 
pydub==0.25.1 
python-dotenv==1.0.0
boto3==1.37.27 
schedule==1.2.2